% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_mtcars-data.R
\docType{data}
\name{big_mtcars}
\alias{big_mtcars}
\title{This dataset is from the EPA. The full data dictionary can be found at fueleconomy.gov.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 41804 rows and 83 columns.
}
\source{
Tidytuesday, \url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2019/2019-10-15}
}
\usage{
data(big_mtcars)
}
\description{
This dataset is from the EPA. The full data dictionary can be found at fueleconomy.gov.
}
\examples{
data(big_mtcars)
}
\keyword{datasets}
