% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_independence.R
\name{chi_square_independence}
\alias{chi_square_independence}
\title{Performs the Chi-square test of independence}
\usage{
chi_square_independence(dat)
}
\arguments{
\item{dat}{numeric, matrix of data, at least 2 by 2}
}
\value{
numeric, a combined vector of length 2, the first element is the uncorrected chi-square statistic, the second is the p-value of the test
}
\description{
Performs the Chi-square test of independence on a input R by C matrix
}
\details{
This function takes a input 2 dimensional matrix (without margins), computes the table margins and the chi-square statistic, and finally calculates the p-value
}
\examples{
chi_square_independence(dat= matrix(c(5,15,25,3), ncol=2,nrow=2) )

}
