% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictactoe.R
\name{tictactoe}
\alias{tictactoe}
\title{Start a tic-tac-toe game on R console}
\usage{
tictactoe(demo = FALSE)
}
\arguments{
\item{demo}{logical, if set to be true, player input will not be needed and
every mark by the player will be randomly decided}
}
\value{
nothing useful is returned.
}
\description{
The player gets to take the first turn, the computer randomly fills in one
of the empty spots each turn.
}
\details{
Start a tic-tac-toe game on R console. The user can specify the
position to mark with command line input. The player takes the first turn,
and the computer marks one spot randomly after every player turn. The board
is printed to the console.
}
\examples{
tictactoe(demo=TRUE)

}
